Move header serialization into policy
